+++ src/Contract/RegistryInterface.php
@@ -1,33 +1,33 @@
 <?php
 /**
- * The file is part of the annotation.
+ * The file is part of the attribute.
  *
- * (c) alan <alan1766447919@gmail.com>.
+ * (c) alan <anhoder@88.com>.
  *
  * 2020/12/21 1:33 下午
  */
 
-namespace Anhoder\Annotation\Contract;
+namespace Yarfox\Attribute\Contract;
 
-use Anhoder\Annotation\Entity\ClassEntity;
+use Yarfox\Attribute\Entity\ClassEntity;
 
 /**
  * Interface AttributeRegistryInterface
  * The interface of class attribute container.
- * @package Anhoder\Annotation\Contract
+ * @package Yarfox\Attribute\Contract
  * @internal
  */
 interface RegistryInterface
 {
     /**
-     * Get annotations from container.
+     * Get attributes from container.
      * @param string|null $namespace
      * @return \ReflectionAttribute[]
      */
     public function getAttributes(string $namespace = null): array;
 
     /**
-     * Get attribute handler by annotation class name.
+     * Get attribute handler by attribute class name.
      * @param string $attribute
      * @return HandlerInterface|null
      */
@@ -39,12 +39,12 @@ interface RegistryInterface
      * @param ClassEntity $classAttributeEntity
      * @return void
      */
-    public function registerAttribute(string $namespace, string $className, ClassEntity $classAttributeEntity);
+    public function registerAttribute(string $namespace, string $className, ClassEntity $classAttributeEntity): void;
 
     /**
      * @param string $attributeName
      * @param string $attributeHandler
      * @return void
      */
-    public function registerAttributeHandler(string $attributeName, string $attributeHandler);
+    public function registerAttributeHandler(string $attributeName, string $attributeHandler): void;
 }