+++ src/Entity/PropertyEntity.php
@@ -1,16 +1,16 @@
 <?php
 /**
- * The file is part of the annotation.
+ * The file is part of the attribute.
  *
- * (c) alan <alan1766447919@gmail.com>.
+ * (c) alan <anhoder@88.com>.
  *
  * 2020/12/21 11:09 下午
  */
 
-namespace Anhoder\Annotation\Entity;
+namespace Yarfox\Attribute\Entity;
 
-use Anhoder\Annotation\Contract\EntityInterface;
-use Anhoder\Annotation\Exception\ReflectionErrorException;
+use Yarfox\Attribute\Contract\EntityInterface;
+use Yarfox\Attribute\Exception\ReflectionErrorException;
 use Attribute;
 use ReflectionAttribute;
 use ReflectionProperty;
@@ -18,7 +18,7 @@ use Reflector;
 
 /**
  * Class PropertyAttributeEntity
- * @package Anhoder\Annotation\Entity
+ * @package Yarfox\Attribute\Entity
  */
 class PropertyEntity implements EntityInterface
 {