+++ tests/app/AttributeConfig.php
@@ -1,15 +1,15 @@
 <?php
 /**
- * The file is part of the annotation.
+ * The file is part of the attribute.
  *
  * (c) anhoder <anhoder@88.com>.
  *
  * 2021/1/12 8:57 上午
  */
 
-namespace Anhoder\Annotation\Test;
+namespace Yarfox\Attribute\Test;
 
-use Anhoder\Annotation\Contract\ConfigInterface;
+use Yarfox\Attribute\Contract\ConfigInterface;
 use JetBrains\PhpStorm\ArrayShape;
 
 class AttributeConfig implements ConfigInterface