+++ tests/app/Handler/ClassAttributeHandler.php
@@ -1,17 +1,17 @@
 <?php
 /**
- * The file is part of the annotation.
+ * The file is part of the attribute.
  *
  * (c) anhoder <anhoder@88.com>.
  *
  * 2021/1/12 8:52 上午
  */
 
-namespace Anhoder\Annotation\Test\Handler;
+namespace Yarfox\Attribute\Test\Handler;
 
-use Anhoder\Annotation\Attribute\AttributeHandler;
-use Anhoder\Annotation\Handler\AbstractHandler;
-use Anhoder\Annotation\Test\Annotation\ClassAttribute;
+use Yarfox\Attribute\Attribute\AttributeHandler;
+use Yarfox\Attribute\Handler\AbstractHandler;
+use Yarfox\Attribute\Test\Attribute\ClassAttribute;
 
 #[AttributeHandler(ClassAttribute::class)]
 class ClassAttributeHandler extends AbstractHandler
@@ -26,7 +26,7 @@ class ClassAttributeHandler extends AbstractHandler
         return self::$attributes;
     }
 
-    public function handle()
+    public function handle(): void
     {
         $attribute = $this->attribute;
         var_dump($attribute->getTest());